Move renderBufferedImage() to effect/Java2d, add a writer marker class, move Java2dBufferedImageWriter to separate file
